fixup(l1): rebase EIP-7805 onto bal-devnet-6
Two rebase-fix touch-ups after rebasing FOCIL onto bal-devnet-6:

- crates/blockchain/payload.rs: bal-devnet-6 changed `set_bal_index`'s
  parameter from u16 to u32 (PR #6521 / EIP-7928 builder/validator parity).
  Update the cast in `apply_inclusion_list_transactions` accordingly.

- test/tests/blockchain/builder_validator_parity_tests.rs: bal-devnet-6's
  new parity test constructs `BuildPayloadArgs` directly. Now that
  `inclusion_list_transactions` is an unconditional `Option<Vec<Transaction>>`
  field on the struct (so cross-crate constructors don't need to thread the
  feature flag), this caller must initialize it. Set to `None` since the
  parity tests don't exercise FOCIL.

+++ crates/blockchain/payload.rs
@@ -754,7 +754,7 @@ impl Blockchain {
 
             // BAL index + per-tx checkpoint, then record touched addresses.
             #[allow(clippy::cast_possible_truncation)]
-            let tx_index = (context.payload.body.transactions.len() + 1) as u16;
+            let tx_index = (context.payload.body.transactions.len() + 1) as u32;
             context.vm.set_bal_index(tx_index);
             let bal_checkpoint = context
                 .vm
